+++ tests/Classes/Basic.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Parable\Tests\Classes;
+
+class Basic
+{
+    public $value = "new";
+}