+++ tests/Classes/CyclicA.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Parable\Tests\Classes;
+
+class CyclicA
+{
+    public function __construct(\Parable\Tests\Classes\CyclicB $b)
+    {
+    }
+}