+++ tests/Classes/CyclicB.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Parable\Tests\Classes;
+
+class CyclicB
+{
+    public function __construct(\Parable\Tests\Classes\CyclicA $a)
+    {
+    }
+}